views(SpotLight): dissolve CSpotFocus/CSpotLaser onto CGameObject; name class-id 5
CSpotLight::m_98 (the serialized focus slot) is a real CGameObject*, not the
fabricated views it was cast to. Retype m_98 -> CGameObject* and dissolve:
  * CSpotFocus  -> CGameObject: its 9 fabricated virtual slots (s0..s7 + "GetType")
    were CGameObject's real slot-8 GetTypeId (+0x20); its +0x188 the object's
    archive-cue id (m_188). SerializeMove's Read probe is the same
    GetTypeId()==5 + id->object map resolve CPlay::SerializeMove uses.
  * CSpotLaser  -> CGameObject: its "+0x5c/+0x60 move-delta" are m_screenX/m_screenY.
Ufo.cpp's owner-stash becomes cast-free (m_98 = m_object, both CGameObject*).
CSpotTarget stays a flagged @identity-TODO view (the Probe_32ce/QueryAt hit record
is larger than any recovered class; its +0x10 wrapped object IS a real CGameObject).

Name the magic 5: add CLASSID_SERIALREF=5 to LoadableClassId (Loadable.h) - the
serialize-map referent kind both probes check - and use it at both sites
(SpotLightCtor + Play). Value proven by two probe sites; owning class still TBD.
Enum value is matching-neutral (same immediate). All byte-neutral: spotlightctor
held 3/6, play 54/110, ufo 4/5.

Also refresh the auto-generated README + cleanliness baseline (.cpp-local views
484 -> 465; the -3 exact in the overall is unrelated environmental rebuild drift,
MAX % held 100%).

(cherry picked from commit 427457ef951ad95b1581ba11a1086d3ff84d716e)

+++ src/Gruntz/Ufo.cpp
@@ -65,8 +65,8 @@ CUFO::CUFO(CGameObject* obj) : CPathHazard(obj) {
             sl->m_120 = m_object->m_130;
             sub->m_notify(sl);
             // The spotlight's bound logic leaf (CSpotLight): stash the UFO's owner
-            // game-object into its reused +0x98 slot (downcast at the site).
-            ((CSpotLight*)sl->m_7c->m_logic)->m_98 = (i32)m_object;
+            // game-object into its reused +0x98 focus slot (both CGameObject*).
+            ((CSpotLight*)sl->m_7c->m_logic)->m_98 = m_object;
         }
     }
     m_object->m_drawActive = 1;
